BAU: Ensure REVISION file is present in images
The apps rely on REVISION to present the GIT_SHA to the healthcheck
controller. This was supplied through a step before calling docker
build, by echoing the output of git rev-parse --short HEAD into a file.

This commit adds this step into the build script which will ensure the
file is present in all containers built using the new orb job.